+++ filament/src/details/Renderer.h
@@ -33,17 +33,13 @@
 #include <fg/FrameGraphTexture.h>
 
 #include <filament/Renderer.h>
-#include <filament/View.h>
 #include <filament/Viewport.h>
 
 #include <backend/DriverEnums.h>
 #include <backend/Handle.h>
-#include <backend/PresentCallable.h>
 
 #include <utils/compiler.h>
 #include <utils/Allocator.h>
-#include <utils/JobSystem.h>
-#include <utils/Slice.h>
 
 #include <tsl/robin_set.h>
 
@@ -53,12 +49,9 @@ namespace backend {
 class Driver;
 } // namespace backend
 
-class View;
-
 class FEngine;
 class FRenderTarget;
 class FView;
-class ShadowMap;
 
 /*
  * A concrete implementation of the Renderer Interface.
@@ -68,39 +61,42 @@ class FRenderer : public Renderer {
 
 public:
     explicit FRenderer(FEngine& engine);
+
     ~FRenderer() noexcept;
 
-    void init() noexcept;
+    void terminate(FEngine& engine);
 
     FEngine& getEngine() const noexcept { return mEngine; }
 
     math::float4 getShaderUserTime() const { return mShaderUserTime; }
 
-    // do all the work here!
-    void renderJob(ArenaScope& arena, FView& view);
+    void resetUserTime();
 
+    // renders a single standalone view. The view must have a a custom rendertarget.
+    void renderStandaloneView(FView const* view);
+
+    // start a frame
     bool beginFrame(FSwapChain* swapChain, uint64_t vsyncSteadyClockTimeNano);
 
+    // end a frame
+    void endFrame();
+
+    // render a view. must be called between beginFrame/enfFrame.
     void render(FView const* view);
 
+    // read pixel from the current swapchain. must be called between beginFrame/enfFrame.
     void readPixels(uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
             backend::PixelBufferDescriptor&& buffer);
 
-    void copyFrame(FSwapChain* dstSwapChain, Viewport const& dstViewport,
-            Viewport const& srcViewport, CopyFrameFlag flags);
-
-    void endFrame();
-
-    void renderStandaloneView(FView const* view);
-
+    // read pixel from a rendertarget. must be called between beginFrame/enfFrame.
     void readPixels(FRenderTarget* renderTarget,
             uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
             backend::PixelBufferDescriptor&& buffer);
 
-    void resetUserTime();
+    // blits the current swapchain to another one
+    void copyFrame(FSwapChain* dstSwapChain, Viewport const& dstViewport,
+            Viewport const& srcViewport, CopyFrameFlag flags);
 
-    // Clean-up everything, this is typically called when the client calls Engine::destroyRenderer()
-    void terminate(FEngine& engine);
 
     void setDisplayInfo(DisplayInfo const& info) noexcept {
         mDisplayInfo = info;
@@ -131,88 +127,45 @@ public:
 private:
     friend class Renderer;
     using Command = RenderPass::Command;
+    using clock = std::chrono::steady_clock;
+    using Epoch = clock::time_point;
+    using duration = clock::duration;
+
+    void initializeClearFlags();
+
+    backend::TextureFormat getHdrFormat(const FView& view, bool translucent) const noexcept;
+    backend::TextureFormat getLdrFormat(bool translucent) const noexcept;
+
+    Epoch getUserEpoch() const { return mUserEpoch; }
+    duration getUserTime() const noexcept { return clock::now() - getUserEpoch(); }
 
     void getRenderTarget(FView const& view,
             backend::TargetBufferFlags& outAttachementMask,
             backend::Handle<backend::HwRenderTarget>& outTarget) const noexcept;
-
-    void readPixels(backend::Handle<backend::HwRenderTarget> renderTargetHandle,
-            uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
-            backend::PixelBufferDescriptor&& buffer);
-
-    void renderInternal(FView const* view);
-
-    struct ColorPassConfig {
-        // Rendering viewport width (e.g. scaled down viewport from dynamic resolution)
-        uint32_t width;
-        // Rendering viewport height (e.g. scaled down viewport from dynamic resolution)
-        uint32_t height;
-        // Rendering offset within the viewport (e.g. non-zero when we have guard bands)
-        uint32_t xoffset;
-        // Rendering offset within the viewport (e.g. non-zero when we have guard bands)
-        uint32_t yoffset;
-        // dynamic resolution scale
-        math::float2 scale;
-        // HDR format
-        backend::TextureFormat hdrFormat;
-        // MSAA sample count
-        uint8_t msaa;
-        // Clear flags
-        backend::TargetBufferFlags clearFlags;
-        // Clear color
-        math::float4 clearColor = {};
-        // Lod offset for the SSR passes
-        float ssrLodOffset;
-        // Contact shadow enabled?
-        bool hasContactShadows;
-    };
-
-    FrameGraphId<FrameGraphTexture> colorPass(FrameGraph& fg, const char* name,
-            FrameGraphTexture::Descriptor const& colorBufferDesc,
-            ColorPassConfig const& config,
-            PostProcessManager::ColorGradingConfig colorGradingConfig,
-            RenderPass::Executor const& passExecutor, FView const& view) const noexcept;
-
-    FrameGraphId<FrameGraphTexture> refractionPass(FrameGraph& fg,
-            ColorPassConfig config,
-            PostProcessManager::ScreenSpaceRefConfig const& ssrConfig,
-            PostProcessManager::ColorGradingConfig colorGradingConfig,
-            RenderPass const& pass, FView const& view) const noexcept;
 
     void recordHighWatermark(size_t watermark) noexcept {
         mCommandsHighWatermark = std::max(mCommandsHighWatermark, watermark);
     }
 
     size_t getCommandsHighWatermark() const noexcept {
-        return mCommandsHighWatermark * sizeof(RenderPass::Command);
+        return mCommandsHighWatermark;
     }
 
-    backend::TextureFormat getHdrFormat(const View& view, bool translucent) const noexcept;
-    backend::TextureFormat getLdrFormat(bool translucent) const noexcept;
-
-    void initializeClearFlags();
-
-    using clock = std::chrono::steady_clock;
-    using Epoch = clock::time_point;
-    using duration = clock::duration;
-
-    Epoch getUserEpoch() const { return mUserEpoch; }
-    duration getUserTime() const noexcept {
-        return clock::now() - getUserEpoch();
-    }
+    void renderInternal(FView const* view);
+    void renderJob(ArenaScope& arena, FView& view);
 
     // keep a reference to our engine
     FEngine& mEngine;
     FrameSkipper mFrameSkipper;
-    backend::Handle<backend::HwRenderTarget> mRenderTarget;
+    backend::Handle<backend::HwRenderTarget> mRenderTargetHandle;
     FSwapChain* mSwapChain = nullptr;
     size_t mCommandsHighWatermark = 0;
     uint32_t mFrameId = 0;
     uint32_t mViewRenderedCount = 0;
     FrameInfoManager mFrameInfoManager;
-    backend::TextureFormat mHdrTranslucent{};
-    backend::TextureFormat mHdrQualityMedium{};
-    backend::TextureFormat mHdrQualityHigh{};
+    backend::TextureFormat mHdrTranslucent;
+    backend::TextureFormat mHdrQualityMedium;
+    backend::TextureFormat mHdrQualityHigh;
     bool mIsRGB8Supported : 1;
     Epoch mUserEpoch;
     math::float4 mShaderUserTime{};
